Added the Integration tests for the TableStorageEntityStore.
Also changed the TableStorageId to a struct and overrode equality methods.  Seemed logical since should never be null and inline memory allocation seemed ok for the intended usage.  I do not think it would cause any memory issues and maybe save some memory utilization if implemented properly, might need a little more work to optimize in the implementation but should be fine for now.

Finally added the TableStorageEntity wrapper class to codify the TableStorageId implementation for the entities of this data source.